+++ src/main/java/com/clubber/ClubberServer/domain/review/repository/ReviewCustomRepository.java
@@ -1,5 +1,6 @@
 package com.clubber.ClubberServer.domain.review.repository;
 
+import com.clubber.ClubberServer.domain.club.domain.Club;
 import com.clubber.ClubberServer.domain.review.domain.Review;
 import com.clubber.ClubberServer.domain.user.domain.User;
 import java.util.List;
@@ -9,4 +10,6 @@ import java.util.List;
 public interface ReviewCustomRepository {
 
     List<Review> queryReviewByUserOrderByIdDesc(User user);
+
+    List<Review> queryReviewByClub(Club club);
 }
