+++ src/main/java/com/clubber/ClubberServer/domain/review/repository/ReviewKeywordCustomRepository.java
@@ -16,6 +16,4 @@ public interface ReviewKeywordCustomRepository {
 
     List<ReviewKeyword> queryReviewKeywordByUserId(Long userId);
 
-    List<Review> queryReviewByClub(Club club);
-
 }
